requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
werkzeug==3.0.3
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3
